# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Crypto Backtester API
  version: 0.8.0
  description: |
    Chart client API — unified chart-data, historical candles, indicator
    compute, structured symbols, users, watchlists, replay (WebSocket v2),
    backtest HTTP, screener scan, AI NL→DSL, and live candle WS.

    **Base path (REST):** `/api/v1`

    **Authentication:** none. Optional `user_id` on create bodies is metadata only.
    Resources are open by UUID. `/auth/*` and `/ws/tickets` have been removed.

    **Timestamps:** Candle `time`, replay `start`/`end`/`cursor`, backtest
    windows, scan windows, and indicator `from`/`to` use **Unix seconds UTC**.

    **Errors (REST):** JSON envelope `{"error": {"code": "...", "message": "..."}}`.

    **WebSocket:** Documented under `x-websocket` (not part of OpenAPI path model).
  contact:
    name: crypto-backtester
  license:
    name: Unspecified
servers:
  - url: http://localhost:8000
    description: Local uvicorn (`python -m api` or `uvicorn api.main:app`)
...
paths:
...
  /api/v1/users:
    post:
      tags: [users]
      summary: Create user (removed)
      description: >
        Removed (BE-L2-019). Use `POST /auth/register` which returns a JWT and
        provisions the default watchlist in the same transaction.
      operationId: createUser
      deprecated: true
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "410":
          description: Endpoint removed; use `POST /auth/register`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                gone:
                  value:
                    error:
                      code: GONE
                      message: "POST /users is disabled; use POST /auth/register"
    get:
      tags: [users]
      summary: List users (removed)
      description: >
        Removed (BE-016). Use `GET /auth/me` for the authenticated user.
      operationId: listUsers
      deprecated: true
      security:
        - bearerAuth: []
      responses:
        "410":
          description: Enumeration removed; use `GET /auth/me`
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                gone:
                  value:
                    error:
                      code: GONE
                      message: "GET /users enumeration is disabled; use GET /auth/me"
  /api/v1/users/{user_id}:
    get:
      tags: [users]
      summary: Get user (owner-only)
      operationId: getUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [users]
      summary: Update user (owner-only)
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
    delete:
      tags: [users]
      summary: Delete user (owner-only)
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "204":
          description: User and cascaded rows deleted (BE-L2-001)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/backtest:
    post:
      tags: [backtest]
      summary: Run a backtest
      description: >
        Synchronously execute the Phase 3 backtest engine and persist the run
        against the JWT subject. Provide exactly one of `strategy_name` or
        inline `strategy`. Does not write equity PNG or trades CSV.
      operationId: createBacktest
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestCreateRequest"
      responses:
        "201":
          description: Persisted backtest run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRunResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/backtest/{run_id}:
    get:
      tags: [backtest]
      summary: Get backtest run (owner-only)
      operationId: getBacktest
      security:
        - bearerAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Persisted run summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestRunResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/backtest/{run_id}/trades:
    get:
      tags: [backtest]
      summary: Get backtest trade log (owner-only)
      operationId: getBacktestTrades
      security:
        - bearerAuth: []
      parameters:
        - name: run_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Full round-trip trade log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BacktestTradesResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/scan:
    post:
      tags: [scan]
      summary: Run a multi-symbol screener scan
      description: >
        Synchronously evaluate a condition tree across symbols and timeframes.
        Defaults to active catalog symbols when `symbols` is omitted. Persists
        to `app.scan_runs` when `persist` is true (default). Alert delivery is
        console/log only (email/webhook deferred).
      operationId: createScan
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanCreateRequest"
      responses:
        "201":
          description: Scan completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRunResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/scan/{scan_id}:
    get:
      tags: [scan]
      summary: Get persisted scan run (owner-only)
      operationId: getScan
      security:
        - bearerAuth: []
      parameters:
        - name: scan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Persisted scan run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanRunResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/ai/translate:
    post:
      tags: [ai]
      summary: Translate natural language into a Trading DSL strategy
      description: >
        LLM returns a validated Phase 9 strategy JSON, or `needs_clarification`
        with questions when the request is ambiguous (D-112–D-114). Per-user
        AI RPM limiter enforced.
      operationId: aiTranslate
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiTranslateRequest"
      responses:
        "200":
          description: Strategy or clarification questions
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AiTranslateOkResponse"
                  - $ref: "#/components/schemas/AiTranslateClarifyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          description: Upstream LLM provider failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/ai/clarify:
    post:
      tags: [ai]
      summary: Answer clarification questions and continue translation
      operationId: aiClarify
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiClarifyRequest"
      responses:
        "200":
          description: Strategy or further clarification
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AiTranslateOkResponse"
                  - $ref: "#/components/schemas/AiTranslateClarifyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/ai/explain:
    post:
      tags: [ai]
      summary: Explain a DSL strategy in plain English
      description: Template walk of the condition tree (no LLM required).
      operationId: aiExplain
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiExplainRequest"
      responses:
        "200":
          description: English explanation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiExplainResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/replay/sessions:
    post:
      tags: [replay]
      summary: Create replay session
      operationId: createReplaySession
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaySessionCreate"
      responses:
        "201":
          description: Session id and WebSocket URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplaySessionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/replay/sessions/{session_id}:
    get:
      tags: [replay]
      summary: Get replay session state (owner-only)
      operationId: getReplaySession
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Session snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayStateResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [replay]
      summary: Delete replay session (owner-only)
      operationId: deleteReplaySession
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204":
          description: Session torn down
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/auth/register` or `/auth/login`.
  parameters:
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            missing:
              value:
                error:
                  code: UNAUTHORIZED
                  message: "Unauthorized"
            invalid:
              value:
                error:
                  code: INVALID_TOKEN
                  message: "Invalid or expired token"
    Forbidden:
      description: Authenticated user is not the owner of the target resource
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            not_owner:
              value:
                error:
                  code: FORBIDDEN
                  message: "Cannot access another user's resources"
    RateLimited:
      description: Rate limit exceeded (per-user or anonymous limiter)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            ai_rpm:
              value:
                error:
                  code: RATE_LIMITED
                  message: "AI rate limit exceeded (30 requests per minute)"
            register:
              value:
                error:
                  code: RATE_LIMITED
                  message: "Too many registration attempts"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            symbol:
              value:
                error:
                  code: SYMBOL_NOT_FOUND
                  message: "Unknown symbol: FOO/USDT"
            user:
              value:
                error:
                  code: USER_NOT_FOUND
                  message: "Unknown user: ..."
            replay:
              value:
                error:
                  code: REPLAY_NOT_FOUND
                  message: "Unknown replay session: ..."
            scan:
              value:
                error:
                  code: SCAN_NOT_FOUND
                  message: "Scan not found"
            backtest:
              value:
                error:
                  code: BACKTEST_NOT_FOUND
                  message: "Backtest run not found"
            ai_session:
              value:
                error:
                  code: SESSION_NOT_FOUND
                  message: "Unknown or expired session"
    ValidationError:
      description: Validation or business rule failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            timeframe:
              value:
                error:
                  code: INVALID_TIMEFRAME
                  message: "Unsupported timeframe: 2m"
            range:
              value:
                error:
                  code: INVALID_RANGE
                  message: from must be <= to
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorBody"
...
    UserCreate:
      type: object
      required: [name, email]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        email:
          type: string
          minLength: 3
          maxLength: 320
          format: email
...
    UserUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          nullable: true
        email:
          type: string
          minLength: 3
          maxLength: 320
          format: email
          nullable: true
    UserResponse:
      type: object
      required: [id, name, email, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ReplaySessionCreate:
      type: object
      required: [symbol, timeframe, start]
      properties:
        symbol:
          type: string
          example: BTC/USDT
        timeframe:
          type: string
          example: 1h
        start:
          type: integer
          format: int64
          description: Open-ended replay start anchor (unix seconds)
        indicators:
          type: array
          items:
            $ref: "#/components/schemas/IndicatorSpec"
          default: []
        step_timeframe:
          type: string
          nullable: true
          description: Bar step resolution (defaults to `timeframe`)
        speed:
          type: number
          exclusiveMinimum: 0
          default: 1.0
          description: Client playback speed (1x = 1 bar/sec)
        autoplay:
          type: boolean
          default: false
      description: >
        `user_id` was removed in BE-L2-006 — the JWT subject owns every
        session. Sending it is ignored today and may return 422 in a future
        release.
    ReplaySessionResponse:
      type: object
      required: [session_id, ws_url]
      properties:
        session_id:
          type: string
          format: uuid
        ws_url:
          type: string
          example: /ws/replay/550e8400-e29b-41d4-a716-446655440000
    ReplayStateResponse:
      type: object
      required:
        - session_id
        - symbol
        - timeframe
        - step_timeframe
        - start
        - state
        - speed
        - bar_index
        - queueRemaining
        - indicators
      properties:
        session_id:
          type: string
          format: uuid
        symbol:
          type: string
        timeframe:
          type: string
        step_timeframe:
          type: string
        start:
          type: integer
          format: int64
          description: Replay start anchor (startAnchor)
        latest_available:
          type: integer
          format: int64
          nullable: true
          description: Latest stored candle for symbol+timeframe (live-updated)
        cursor:
          type: integer
          format: int64
          nullable: true
          description: Current bar time (unix seconds)
        state:
          type: string
          enum: [idle, playing, paused, completed]
        speed:
          type: number
        bar_index:
          type: integer
          description: Revealed bars in current trail window (barIndex)
        queueRemaining:
          type: integer
          description: Bars remaining in server prefetch queue (not user progress)
        indicators:
          type: array
          items:
            $ref: "#/components/schemas/IndicatorSpec"

    # --- WebSocket client commands ---
...
    BacktestCreateRequest:
      type: object
      required: [symbol, timeframe, start, end]
      properties:
        symbol:
          type: string
        timeframe:
          type: string
        start:
          type: integer
          format: int64
        end:
          type: integer
          format: int64
        initial_capital:
          type: number
          format: double
        strategy_name:
          type: string
          description: XOR with strategy — catalog name from config.yaml
        strategy:
          type: object
          additionalProperties: true
          description: XOR with strategy_name — inline YAML-compatible strategy
        backtest:
          $ref: "#/components/schemas/BacktestParamsBody"
        user_id:
          type: string
          format: uuid
          nullable: true
    ScanCreateRequest:
      type: object
      required: [timeframes, start, end, condition]
      properties:
        timeframes:
          type: array
          items:
            type: string
          minItems: 1
        start:
          type: integer
          format: int64
        end:
          type: integer
          format: int64
        condition:
          type: object
          additionalProperties: true
          description: Signal condition tree (all/any/not + indicator/smc leaves)
        symbols:
          type: array
          items:
            type: string
          nullable: true
          description: Optional explicit list; defaults to active catalog
        alert_trigger:
          type: string
          enum: [edge, level]
          default: edge
        persist:
          type: boolean
          default: true
...
    ScanRunResponse:
      type: object
      required:
        - timeframes
        - symbols
        - start
        - end
        - alert_trigger
        - matches
        - alert_count
        - duration_ms
        - scanned_pairs
        - persisted
      properties:
        scan_id:
          type: string
          format: uuid
          nullable: true
        timeframes:
          type: array
          items:
            type: string
        symbols:
          type: array
          items:
            type: string
        start:
          type: integer
          format: int64
        end:
          type: integer
          format: int64
        alert_trigger:
          type: string
          enum: [edge, level]
        matches:
          type: array
          items:
            $ref: "#/components/schemas/ScanMatchResponse"
        alert_count:
          type: integer
        duration_ms:
          type: integer
        scanned_pairs:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ScanErrorResponse"
        persisted:
          type: boolean
        persist_error:
          type: string
          nullable: true
          description: >
            Non-null when `persist=true` and the persist step failed
            (BE-L2-014). Compute results are still returned; caller can
            distinguish infra failure from an intentional `persist=false`.
          example: PERSIST_FAILED
...
    BacktestRunResponse:
      type: object
      required:
        - run_id
        - symbol
        - timeframe
        - start
        - end
        - initial_capital
        - status
        - metrics
        - equity
        - signals
        - trades
        - created_at
      properties:
        run_id:
          type: string
          format: uuid
        symbol:
          type: string
        timeframe:
          type: string
        start:
          type: integer
          format: int64
        end:
          type: integer
          format: int64
        initial_capital:
          type: number
        strategy_name:
          type: string
          nullable: true
        status:
          type: string
        metrics:
          $ref: "#/components/schemas/BacktestMetricsResponse"
        equity:
          type: array
          items:
            $ref: "#/components/schemas/EquityPoint"
        signals:
          type: array
          items:
            $ref: "#/components/schemas/ChartSignal"
        trades:
          type: array
          items:
            $ref: "#/components/schemas/ChartTrade"
        created_at:
          type: string
          format: date-time
...
    EquityPoint:
      type: object
      required: [time, value]
      properties:
        time:
          type: integer
          format: int64
        value:
          type: number
...
    BacktestTradesResponse:
      type: object
      required: [run_id, trades]
      properties:
        run_id:
          type: string
          format: uuid
        trades:
          type: array
          items:
            $ref: "#/components/schemas/TradeDetail"
...
    AiTranslateRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
          minLength: 1
          description: Plain-English strategy description
    AiClarifyRequest:
      type: object
      required: [session_id, answers]
      properties:
        session_id:
          type: string
        answers:
          type: object
          additionalProperties:
            type: string
          description: Map of clarification question id → answer
    AiExplainRequest:
      type: object
      required: [strategy]
      properties:
        strategy:
          type: object
          additionalProperties: true
          description: Phase 9 Trading DSL strategy document
...
    AiTranslateOkResponse:
      type: object
      required: [status, strategy]
      properties:
        status:
          type: string
          enum: [ok]
        strategy:
          type: object
          additionalProperties: true
        explanation:
          type: string
    AiTranslateClarifyResponse:
      type: object
      required: [status, session_id, questions]
      properties:
        status:
          type: string
          enum: [needs_clarification]
        session_id:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/AiClarificationQuestion"
    AiExplainResponse:
      type: object
      required: [explanation]
      properties:
        explanation:
          type: string